Restructing code to use encypt and decypt function in other sdk's
